+++ CHANGELOG.md
@@ -1,5 +1,52 @@
 # Changelog
 
+## 0.3.2 - 2026-05-01
+
+This release adds the `WithDatabase` option, which controls the location of the test case database:
+
+```go
+hegel.Test(t, func(ht *hegel.T) {
+    ...
+}, hegel.WithDatabase(hegel.Database("my_custom_directory")))
+
+// disable the database
+hegel.Test(t, func(ht *hegel.T) {
+    ...
+}, hegel.WithDatabase(hegel.DatabaseDisabled()))
+```
+
+This release also adds the `WithDerandomize` option, which can be set to make the test run deterministically:
+
+```go
+hegel.Test(t, func(ht *hegel.T) {
+    ...
+}, hegel.WithDerandomize(true))
+```
+
+## 0.3.1 - 2026-04-30
+
+Internal refactor.
+
+## 0.3.0 - 2026-04-30
+
+This release removes `hegel.Case` in favor of a new `hegel.Test`. `hegel.Test` is now the recommended way to write Hegel tests.
+
+```go
+// before
+func TestA(t *testing.T) {
+	t.Run("test_name", hegel.Case(func(ht *hegel.T) {
+		hegel.Draw(ht, hegel.Integers(-1000, 1000))
+	}))
+}
+
+// after
+func TestA(t *testing.T) {
+	hegel.Test(t, func(ht *hegel.T) {
+		hegel.Draw(ht, hegel.Integers(-1000, 1000))
+	})
+}
+```
+
 ## 0.2.1 - 2026-04-29
 
 Internal refactor of `oneOf`.